# openapi.yaml
openapi: 3.0.0
info:
  title: SignalZero Kernel API
  version: 2.0.0
  description: API for the SignalZero Recursive Symbolic Kernel.
servers:
  - url: http://localhost:3000/api
paths:
...
  /chat:
    post:
      summary: Send a message to the chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
                  toolCalls:
                    type: array
                    items:
                      type: object
...
  /system/prompt:
    get:
      summary: Get the current system prompt
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type: string
    post:
      summary: Set a new system prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /index/reindex:
    post:
      summary: Reindex all symbols
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                includeDisabled:
                  type: boolean
                  description: Include disabled domains in the reindex process
      responses:
        '200':
          description: Reindex operation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReindexResult'
...
  /domains/{id}/exists:
    get:
      summary: Check if a domain exists
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
  /domains/{id}/enabled:
    get:
      summary: Check if a domain is enabled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
  /domains/{id}/toggle:
    post:
      summary: Toggle a domain's enabled status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  domainId:
                    type: string
                  enabled:
                    type: boolean
  /domains/{id}:
    patch:
      summary: Update domain metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                invariants:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
    delete:
      summary: Delete a domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...
  /symbols/search:
    get:
      summary: Search for symbols
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VectorSearchResult'
  /symbols/refactor:
    post:
      summary: Refactor symbols
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                updates:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Successful response
  /symbols/compress:
    post:
      summary: Compress symbols
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newSymbol:
                  type: object
                oldIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Successful response
  /symbols/{id}:
    get:
      summary: Get a symbol by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolDef'
  /domains/{id}/symbols:
    get:
      summary: Get all symbols in a domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SymbolDef'
    post:
      summary: Upsert a symbol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SymbolDef'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  id:
                    type: string
  /domains/{id}/symbols/bulk:
    post:
      summary: Bulk upsert symbols
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SymbolDef'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  count:
                    type: integer
  /domains/{domainId}/symbols/{symbolId}:
    delete:
      summary: Delete a symbol
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: string
        - name: symbolId
          in: path
          required: true
          schema:
            type: string
        - name: cascade
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /domains/{domainId}/symbols/rename:
    post:
      summary: Rename a symbol
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldId:
                  type: string
                newId:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /tests/sets:
    get:
      summary: List test sets
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestSet'
    post:
      summary: Create/Update a test set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestSet'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /tests:
    post:
      summary: Add a test case to a test set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                testSetId:
                  type: string
                name:
                  type: string
                prompt:
                  type: string
                expectedActivations:
                  type: array
                  items:
                    type: string
              required:
                - testSetId
                - name
                - prompt
                - expectedActivations
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /tests/{testSetId}/{testId}:
    delete:
      summary: Delete a test case from a test set
      parameters:
        - name: testSetId
          in: path
          required: true
          schema:
            type: string
        - name: testId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /tests/sets/{id}:
    delete:
      summary: Delete a test set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /tests/runs:
    get:
      summary: List test runs
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestRun'
    post:
      summary: Start a test run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                testSetId:
                  type: string
                compareWithBaseModel:
                  type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  runId:
                    type: string
  /tests/runs/{id}:
    get:
      summary: Get test run details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRun'
  /project/export:
    post:
      summary: Export the project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                meta:
                  $ref: '#/components/schemas/ProjectMeta'
      responses:
        '200':
          description: Successful response
          content:
            application/zip: {}
...
components:
  schemas:
...
    TestSet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        tests:
          type: array
          items:
            $ref: '#/components/schemas/TestCase'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
...
    IndexQueueState:
      type: object
      properties:
        pending:
          type: integer
        total:
          type: integer
        running:
          type: boolean
    ReindexResult:
      type: object
      properties:
        status:
          type: string
          enum:
            - completed
            - already-running
        resetPerformed:
          type: boolean
        indexedCount:
          type: integer
        totalSymbols:
          type: integer
        failedIds:
          type: array
          items:
            type: string
        lastReindexAt:
          type: string
          format: date-time
        queue:
          $ref: '#/components/schemas/IndexQueueState'
...
    VectorSearchResult:
      type: object
      properties:
        id:
          type: string
        score:
          type: number
        metadata:
          type: object
        document:
          type: string
...